+++ src/mesh/mesh.cpp
@@ -98,7 +98,7 @@ Mesh::Mesh(ParameterInput *pin, ApplicationInput *app_in, Properties_t &properti
       lb_automatic_(),
       lb_manual_(), MeshGenerator_{nullptr, UniformMeshGeneratorX1,
                                    UniformMeshGeneratorX2, UniformMeshGeneratorX3},
-      BoundaryFunction_{nullptr, nullptr, nullptr, nullptr, nullptr, nullptr}, AMRFlag_{},
+      MeshBndryFnctn{nullptr, nullptr, nullptr, nullptr, nullptr, nullptr}, AMRFlag_{},
       UserSourceTerm_{}, UserTimeStep_{} {
   std::stringstream msg;
   RegionSize block_size;
@@ -199,6 +199,8 @@ Mesh::Mesh(ParameterInput *pin, ApplicationInput *app_in, Properties_t &properti
         << std::endl;
     PARTHENON_FAIL(msg);
   }
+
+  EnrollBndryFncts_(app_in);
 
   // read and set MeshBlock parameters
   block_size.x1rat = mesh_size.x1rat;
@@ -537,7 +539,7 @@ Mesh::Mesh(ParameterInput *pin, ApplicationInput *app_in, RestartReader &rr,
       lb_automatic_(),
       lb_manual_(), MeshGenerator_{nullptr, UniformMeshGeneratorX1,
                                    UniformMeshGeneratorX2, UniformMeshGeneratorX3},
-      BoundaryFunction_{nullptr, nullptr, nullptr, nullptr, nullptr, nullptr}, AMRFlag_{},
+      MeshBndryFnctn{nullptr, nullptr, nullptr, nullptr, nullptr, nullptr}, AMRFlag_{},
       UserSourceTerm_{}, UserTimeStep_{} {
   std::stringstream msg;
   RegionSize block_size;
@@ -584,6 +586,7 @@ Mesh::Mesh(ParameterInput *pin, ApplicationInput *app_in, RestartReader &rr,
   if (app_in->UserWorkAfterLoop != nullptr) {
     UserWorkAfterLoop = app_in->UserWorkAfterLoop;
   }
+  EnrollBndryFncts_(app_in);
 
   std::vector<Real> bounds = rr.ReadAttr1DReal("Mesh", "bounds");
   mesh_size.x1min = bounds[0];
@@ -908,6 +911,42 @@ void Mesh::OutputMeshStructure(int ndim) {
 }
 
 //----------------------------------------------------------------------------------------
+//  Enroll user-defined functions for boundary conditions
+void Mesh::EnrollBndryFncts_(ApplicationInput *app_in) {
+  static const BValFunc outflow[6] = {
+      BoundaryFunction::OutflowInnerX1, BoundaryFunction::OutflowOuterX1,
+      BoundaryFunction::OutflowInnerX2, BoundaryFunction::OutflowOuterX2,
+      BoundaryFunction::OutflowInnerX3, BoundaryFunction::OutflowOuterX3};
+  static const BValFunc reflect[6] = {
+      BoundaryFunction::ReflectInnerX1, BoundaryFunction::ReflectOuterX1,
+      BoundaryFunction::ReflectInnerX2, BoundaryFunction::ReflectOuterX2,
+      BoundaryFunction::ReflectInnerX3, BoundaryFunction::ReflectOuterX3};
+
+  for (int f = 0; f < BOUNDARY_NFACES; f++) {
+    switch (mesh_bcs[f]) {
+    case BoundaryFlag::reflect:
+      MeshBndryFnctn[f] = reflect[f];
+      break;
+    case BoundaryFlag::outflow:
+      MeshBndryFnctn[f] = outflow[f];
+      break;
+    case BoundaryFlag::user:
+      if (app_in->boundary_conditions[f] != nullptr) {
+        MeshBndryFnctn[f] = app_in->boundary_conditions[f];
+      } else {
+        std::stringstream msg;
+        msg << "A user boundary condition for face " << f
+            << " was requested. but no condition was enrolled." << std::endl;
+        PARTHENON_THROW(msg);
+      }
+      break;
+    default: // periodic/block BCs handled elsewhere.
+      break;
+    }
+  }
+}
+
+//----------------------------------------------------------------------------------------
 //! \fn void Mesh::EnrollUserRefinementCondition(AMRFlagFunc amrflag)
 //  \brief Enroll a user-defined function for checking refinement criteria
 
@@ -965,15 +1004,6 @@ void Mesh::EnrollUserExplicitSourceFunction(SrcTermFunc my_func) {
 
 void Mesh::EnrollUserTimeStepFunction(TimeStepFunc my_func) {
   UserTimeStep_ = my_func;
-  return;
-}
-
-//----------------------------------------------------------------------------------------
-//! \fn void Mesh::EnrollUserMetric(MetricFunc my_func)
-//  \brief Enroll a user-defined metric for arbitrary GR coordinates
-
-void Mesh::EnrollUserMetric(MetricFunc my_func) {
-  UserMetric_ = my_func;
   return;
 }
 
@@ -1050,24 +1080,9 @@ void Mesh::Initialize(int res_flag, ParameterInput *pin, ApplicationInput *app_i
 #pragma omp for
       for (int i = 0; i < nmb; ++i) {
         auto &pmb = block_list[i];
-        auto &pbval = pmb->pbval;
-        if (multilevel) pbval->ProlongateBoundaries(0.0, 0.0);
-        // TODO(JoshuaSBrown): Dead code left in for possible future extraction of
-        // primitives
-        //        int il = pmb->is, iu = pmb->ie, jl = pmb->js, ju = pmb->je, kl =
-        //        pmb->ks,
-        //           ku = pmb->ke;
-        //        if (pbval->nblevel[1][1][0] != -1) il -= NGHOST;
-        //        if (pbval->nblevel[1][1][2] != -1) iu += NGHOST;
-        //        if (pmb->block_size.nx2 > 1) {
-        //          if (pbval->nblevel[1][0][1] != -1) jl -= NGHOST;
-        //          if (pbval->nblevel[1][2][1] != -1) ju += NGHOST;
-        //        }
-        //        if (pmb->block_size.nx3 > 1) {
-        //          if (pbval->nblevel[0][1][1] != -1) kl -= NGHOST;
-        //          if (pbval->nblevel[2][1][1] != -1) ku += NGHOST;
-        //        }
-
+        if (multilevel) {
+          ProlongateBoundaries(pmb->meshblock_data.Get());
+        }
         ApplyBoundaryConditions(pmb->meshblock_data.Get());
         // Call MeshBlockData based FillDerived functions
         Update::FillDerived(pmb->meshblock_data.Get());